Rename servertelemetry to eststelemetry
